updated meh

+++ SHERPA/PerturbativePhysics/Matrix_Element_Handler.C
@@ -509,7 +509,7 @@ int Matrix_Element_Handler::InitializeProcesses(
   p_isr=isr;
   if (!m_gens.InitializeGenerators(p_model,beam,isr)) return false;
   Settings& s = Settings::GetMainSettings();
-  int initonly=s["INIT_ONLY"].Get<size_t>();
+  int initonly=s["INIT_ONLY"].Get<int>();
   if (initonly&4) return 1;
   double rbtime(ATOOLS::rpa->gen.Timer().RealTime());
   double btime(ATOOLS::rpa->gen.Timer().UserTime());
